macos: fix build warnings on implicit casts
The implicit cast of qsize to int discards the sign information. XCode
warns about this, so explicitly cast to (int) to confirm that the cast
is correct.

+++ src/imported/double-slider.cpp
@@ -31,5 +31,5 @@ void DoubleSlider::intValChanged(int val)
 
 void DoubleSlider::setDoubleVal(double val)
 {
-	setValue(lround((val - minVal) / minStep));
+	setValue((int)lround((val - minVal) / minStep));
 }
